fix: Address critical issues #1-8 and add real-time data streaming
Critical Fixes:
- Firmware update: Added size validation, battery check (30% min), extended reboot delay (3s), and flash write verification
- Ring buffer: Fixed integer overflow in count calculations using modulo arithmetic
- Thread safety: Added memory barriers and proper volatile access patterns for ring buffers
- Watchdog: Added watchdog timer to IMU sampling and CSV conversion tasks
- Buffer overflow: Added filename length validation and snprintf checks in logger
- Battery monitoring: Documented millis() wrap-around safety with explicit cast

Real-Time Streaming:
- Implemented WebSocket server on port 81 for real-time data streaming
- Non-intrusive streaming: samples sent after SD write completion
- Client control via 'start'/'stop' WebSocket messages
- JSON format with ADC force calculation and IMU data
- Rate limiting: 20 Hz update rate (50ms) to minimize interference
- Integrated streaming hook in loggerTick() without affecting logging performance

Technical Improvements:
- Added WebSockets library dependency
- Enhanced firmware upload with comprehensive validation
- Improved error handling and logging throughout
- Maintained backward compatibility with all existing features

+++ src/main.cpp
@@ -280,11 +280,13 @@ void loop()
     rtcHandleUpdate();
 
     // ---- Battery monitoring (check every 5 seconds) ----
+    // Note: millis() wraps after ~49 days, but unsigned subtraction handles this correctly
+    // (millis() - lastBatteryCheck) will work correctly even after wrap-around
     static uint32_t lastBatteryCheck = 0;
     static const uint32_t BATTERY_CHECK_INTERVAL_MS = 5000; // Check every 5 seconds
     static NeopixelPattern lastNonBatteryPattern = NEOPIXEL_PATTERN_OFF;
     
-    if (millis() - lastBatteryCheck >= BATTERY_CHECK_INTERVAL_MS)
+    if ((uint32_t)(millis() - lastBatteryCheck) >= BATTERY_CHECK_INTERVAL_MS)
     {
         lastBatteryCheck = millis();
